Feat : add update Comment API

+++ client/src/components/Detail/CommentCard.tsx
@@ -3,6 +3,7 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import React, { Dispatch, SetStateAction, useState } from "react";
 import { postService } from "src/apis/postAPI";
 import { useUserState } from "src/context/userContext";
+import { colors } from "src/style/colors";
 import { UserState } from "src/types/user";
 import { displayedAt } from "src/utils/convertToTIme";
 import { Comment } from "../../types/post";
@@ -15,12 +16,25 @@ const userImgSrc = "https://morethanmin-remotto.herokuapp.com/images/default-use
 
 function CommentCard({ comment }: Props) {
   const [user] = useUserState() as [UserState, Dispatch<SetStateAction<UserState>>];
+  const [isEdit, setIsEdit] = useState(false);
   const [content, setContent] = useState(comment.content);
+
   const queryClient = useQueryClient();
+
   const { mutate: deleteComment } = useMutation(
     (commentId: number) => postService.deleteComment(commentId),
     {
       onSuccess: () => {
+        queryClient.invalidateQueries(["posts", comment.postId]);
+      },
+    }
+  );
+
+  const { mutate: updateComment } = useMutation(
+    (commentId: number) => postService.updateComment(commentId, { content }),
+    {
+      onSuccess: () => {
+        setIsEdit(false);
         queryClient.invalidateQueries(["posts", comment.postId]);
       },
     }
@@ -41,10 +55,18 @@ function CommentCard({ comment }: Props) {
         <ContentEdit>
           {user.userId === comment.userId && (
             <>
-              <DeleteAndRevise>수정</DeleteAndRevise>
               <DeleteAndRevise
                 onClick={() => {
-                  deleteComment(comment.id);
+                  setIsEdit((pre) => !pre);
+                }}
+              >
+                수정
+              </DeleteAndRevise>
+              <DeleteAndRevise
+                onClick={() => {
+                  if (window.confirm("정말 삭제하시겠습니까?")) {
+                    deleteComment(comment.id);
+                  }
                 }}
               >
                 삭제
@@ -54,8 +76,24 @@ function CommentCard({ comment }: Props) {
         </ContentEdit>
       </User>
       <Content>
-        <CommentInput value={content} disabled={true}></CommentInput>
+        <CommentInput
+          value={content}
+          disabled={!isEdit}
+          onChange={(e) => {
+            setContent(e.target.value);
+          }}
+          isEdit={isEdit}
+        />
       </Content>
+      {isEdit && (
+        <CommentReviseBtn
+          onClick={() => {
+            updateComment(comment.id);
+          }}
+        >
+          수정완료
+        </CommentReviseBtn>
+      )}
     </Container>
   );
 }
@@ -66,6 +104,7 @@ const Container = styled.div`
   width: 100%;
   background-color: white;
   padding: 10px;
+  border: solid 1px lightgray;
 `;
 
 const User = styled.div`
@@ -118,9 +157,24 @@ const Content = styled.div`
   border-radius: 10px;
 `;
 
-const CommentInput = styled.input`
-  background-color: #eeeeee;
+const CommentInput = styled.input<{ isEdit: boolean }>`
   border: none;
   width: 100%;
-  height: 30px;
+  height: 40px;
+  background-color: ${({ isEdit }) => (isEdit ? "white" : "#eeeeee")};
+  border-radius: 5px;
+  padding: 10px;
+
+  :focus {
+    outline: none;
+    border: solid 1px ${colors.mainColor};
+  }
+`;
+
+const CommentReviseBtn = styled.p`
+  margin-top: 10px;
+  font-size: 12px;
+  margin-left: 5px;
+  color: #0095f6;
+  cursor: pointer;
 `;
